Refactor maintenance form and export utils
